Take the client address the proxy observed, not the one the caller claimed
Fiber's c.IP() returns the first valid address in PROXY_HEADER, reading left to
right. Every common proxy appends to X-Forwarded-For rather than replacing it --
an ALB does, and so does nginx's $proxy_add_x_forwarded_for -- so the leftmost
field is written by the caller, and the caller is exactly who the limits are
meant to apply to.

Keying the login backoff on the client address made that worse rather than
better. A caller sending an admin's address was filed as that admin and could
lock them out of the panel, which is the denial of service this branch just
closed wearing the fix's clothes. A caller sending a different value per request
got a fresh rate-limit bucket and a fresh backoff row every time and never
reached any threshold; the account-wide lock used to backstop that and no longer
does.

normalizeClientIP walks the header from the right and stops at the first field
that is not itself a trusted proxy: everything right of it was written by
infrastructure, everything left could have been written by anyone. It rewrites
the header to that one value ahead of every limiter, so c.IP() answers correctly
for the limiters, the backoff and the audit log without each call site choosing.

Measured against a server with PROXY_HEADER=X-Forwarded-For and
TRUSTED_PROXIES=127.0.0.1, reading back the row the backoff writes:

  203.0.113.9, 198.51.100.50                          -> 198.51.100.50
  203.0.113.9, 127.0.0.1, 127.0.0.1, 198.51.100.50    -> 198.51.100.50
  not-an-ip, 203.0.113.9, 198.51.100.50               -> 198.51.100.50
  header absent                                        -> the socket peer
  127.0.0.1, 127.0.0.1                                 -> the socket peer

Two test files change only to pass the widened Setup signature. No assertions
were touched.

Also fixes two stale numbers the retune left behind, in the ARCHITECTURE table
and the shipped README, and widens the decay assertion to uint64: uint is the
platform word, so on a 32-bit build the correct configuration was the one that
failed to compile.

+++ apps/api/internal/service/auth/auth.go
@@ -84,14 +84,16 @@ const (
 
 // This is an assertion, not a value: it fails the build if loginFailureDecay is ever raised to or
 // past maxLockDuration. The subtraction is a compile-time constant, and a negative one does not
-// convert to uint, so the inversion cannot land quietly. It is written as a constant rather than
+// convert to an unsigned type, so the inversion cannot land quietly. uint64 rather than uint,
+// because uint is the platform word: on a 32-bit build the correct configuration would be the one
+// that failed to compile, since a 60-minute duration in nanoseconds does not fit in 32 bits. It is written as a constant rather than
 // as a test because the property is about two constants and nothing else, and a build that cannot
 // produce the inversion is stronger than a test run that catches it.
 //
 // The extra nanosecond makes equal windows fail too: a decay exactly as long as the longest lock
 // still leaves the count one nanosecond short of resetting, which is the same trap one minute
 // wider would be.
-const _ = uint(maxLockDuration - loginFailureDecay - time.Nanosecond)
+const _ = uint64(maxLockDuration - loginFailureDecay - time.Nanosecond)
 
 // lockDuration is the backoff curve: nothing for the first four failures, then doubling from one
 // minute, capped. failures is the count AFTER the failure being recorded.
